yolo-service/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
pillow
pandas
python-multipart